+++ src/coreclr/jit/importercalls.cpp
@@ -3350,7 +3350,7 @@ GenTree* Compiler::impIntrinsic(CORINFO_CLASS_HANDLE    clsHnd,
         return node;
     }
 
-    if (ni == NI_System_Runtime_CompilerServices_RuntimeHelpers_Await)
+    if (ni == NI_System_Runtime_CompilerServices_AsyncHelpers_Await)
     {
         // These are marked intrinsics simply to match them by name in
         // the Await pattern optimization. Make sure we keep pIntrinsicName assigned
@@ -11018,10 +11018,7 @@ NamedIntrinsic Compiler::lookupNamedIntrinsic(CORINFO_METHOD_HANDLE method)
                             {
                                 result = NI_System_Runtime_CompilerServices_RuntimeHelpers_GetMethodTable;
                             }
-                            else if (strcmp(methodName, "Await") == 0)
-                            {
-                                result = NI_System_Runtime_CompilerServices_RuntimeHelpers_Await;
-                            }
+
                             else if (strcmp(methodName, "get_RuntimeAsyncViaJitGeneratedStateMachines") == 0)
                             {
                                 result =
@@ -11033,6 +11030,10 @@ NamedIntrinsic Compiler::lookupNamedIntrinsic(CORINFO_METHOD_HANDLE method)
                             if (strcmp(methodName, "AsyncSuspend") == 0)
                             {
                                 result = NI_System_Runtime_CompilerServices_AsyncHelpers_AsyncSuspend;
+                            }
+                            else if (strcmp(methodName, "Await") == 0)
+                            {
+                                result = NI_System_Runtime_CompilerServices_AsyncHelpers_Await;
                             }
                         }
                         else if (strcmp(className, "StaticsHelpers") == 0)
