+++ src/coreclr/jit/namedintrinsiclist.h
@@ -119,10 +119,10 @@ enum NamedIntrinsic : unsigned short
     NI_System_Runtime_CompilerServices_RuntimeHelpers_IsKnownConstant,
     NI_System_Runtime_CompilerServices_RuntimeHelpers_IsReferenceOrContainsReferences,
     NI_System_Runtime_CompilerServices_RuntimeHelpers_GetMethodTable,
-    NI_System_Runtime_CompilerServices_RuntimeHelpers_Await,
     NI_System_Runtime_CompilerServices_RuntimeHelpers_get_RuntimeAsyncViaJitGeneratedStateMachines,
 
     NI_System_Runtime_CompilerServices_AsyncHelpers_AsyncSuspend,
+    NI_System_Runtime_CompilerServices_AsyncHelpers_Await,
 
     NI_System_Runtime_CompilerServices_StaticsHelpers_VolatileReadAsByref,
