+++ src/libraries/System.Private.CoreLib/src/System/Runtime/CompilerServices/RuntimeHelpers.cs
@@ -175,153 +175,5 @@ namespace System.Runtime.CompilerServices
         /// <returns>true if the given type is a reference type or a value type that contains references or by-refs; otherwise, false.</returns>
         [Intrinsic]
         public static bool IsReferenceOrContainsReferences<T>() where T: allows ref struct => IsReferenceOrContainsReferences<T>();
-
-#if !NATIVEAOT
-        [BypassReadyToRun]
-        [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.Async)]
-        public static void AwaitAwaiterFromRuntimeAsync<TAwaiter>(TAwaiter awaiter) where TAwaiter : INotifyCompletion
-        {
-            ref AsyncHelpers.RuntimeAsyncAwaitState state = ref AsyncHelpers.t_runtimeAsyncAwaitState;
-            Continuation? sentinelContinuation = state.SentinelContinuation;
-            if (sentinelContinuation == null)
-                state.SentinelContinuation = sentinelContinuation = new Continuation();
-
-            state.Notifier = awaiter;
-            AsyncHelpers.AsyncSuspend(sentinelContinuation);
-        }
-
-        [BypassReadyToRun]
-        [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.Async)]
-        public static void UnsafeAwaitAwaiterFromRuntimeAsync<TAwaiter>(TAwaiter awaiter) where TAwaiter : ICriticalNotifyCompletion
-        {
-            ref AsyncHelpers.RuntimeAsyncAwaitState state = ref AsyncHelpers.t_runtimeAsyncAwaitState;
-            Continuation? sentinelContinuation = state.SentinelContinuation;
-            if (sentinelContinuation == null)
-                state.SentinelContinuation = sentinelContinuation = new Continuation();
-
-            state.Notifier = awaiter;
-            AsyncHelpers.AsyncSuspend(sentinelContinuation);
-        }
-
-        // Marked intrinsic since JIT recognises the helper by name when doing optimizations.
-        [Intrinsic]
-        [BypassReadyToRun]
-        [MethodImpl(MethodImplOptions.Async)]
-        public static T Await<T>(Task<T> task)
-        {
-            TaskAwaiter<T> awaiter = task.GetAwaiter();
-            if (!awaiter.IsCompleted)
-            {
-                UnsafeAwaitAwaiterFromRuntimeAsync(awaiter);
-            }
-
-            return awaiter.GetResult();
-        }
-
-        // Marked intrinsic since JIT recognises the helper by name when doing optimizations.
-        [Intrinsic]
-        [BypassReadyToRun]
-        [MethodImpl(MethodImplOptions.Async)]
-        public static void Await(Task task)
-        {
-            TaskAwaiter awaiter = task.GetAwaiter();
-            if (!awaiter.IsCompleted)
-            {
-                UnsafeAwaitAwaiterFromRuntimeAsync(awaiter);
-            }
-
-            awaiter.GetResult();
-        }
-
-        // Marked intrinsic since JIT recognises the helper by name when doing optimizations.
-        [Intrinsic]
-        [BypassReadyToRun]
-        [MethodImpl(MethodImplOptions.Async)]
-        public static T Await<T>(ValueTask<T> task)
-        {
-            ValueTaskAwaiter<T> awaiter = task.GetAwaiter();
-            if (!awaiter.IsCompleted)
-            {
-                UnsafeAwaitAwaiterFromRuntimeAsync(awaiter);
-            }
-
-            return awaiter.GetResult();
-        }
-
-        // Marked intrinsic since JIT recognises the helper by name when doing optimizations.
-        [Intrinsic]
-        [BypassReadyToRun]
-        [MethodImpl(MethodImplOptions.Async)]
-        public static void Await(ValueTask task)
-        {
-            ValueTaskAwaiter awaiter = task.GetAwaiter();
-            if (!awaiter.IsCompleted)
-            {
-                UnsafeAwaitAwaiterFromRuntimeAsync(awaiter);
-            }
-
-            awaiter.GetResult();
-        }
-
-        // Marked intrinsic since JIT recognises the helper by name when doing optimizations.
-        [Intrinsic]
-        [BypassReadyToRun]
-        [MethodImpl(MethodImplOptions.Async)]
-        public static void Await(ConfiguredTaskAwaitable configuredAwaitable)
-        {
-            ConfiguredTaskAwaitable.ConfiguredTaskAwaiter awaiter = configuredAwaitable.GetAwaiter();
-            if (!awaiter.IsCompleted)
-            {
-                UnsafeAwaitAwaiterFromRuntimeAsync(awaiter);
-            }
-
-            awaiter.GetResult();
-        }
-
-        // Marked intrinsic since JIT recognises the helper by name when doing optimizations.
-        [Intrinsic]
-        [BypassReadyToRun]
-        [MethodImpl(MethodImplOptions.Async)]
-        public static void Await(ConfiguredValueTaskAwaitable configuredAwaitable)
-        {
-            ConfiguredValueTaskAwaitable.ConfiguredValueTaskAwaiter awaiter = configuredAwaitable.GetAwaiter();
-            if (!awaiter.IsCompleted)
-            {
-                UnsafeAwaitAwaiterFromRuntimeAsync(awaiter);
-            }
-
-            awaiter.GetResult();
-        }
-
-        // Marked intrinsic since JIT recognises the helper by name when doing optimizations.
-        [Intrinsic]
-        [BypassReadyToRun]
-        [MethodImpl(MethodImplOptions.Async)]
-        public static T Await<T>(ConfiguredTaskAwaitable<T> configuredAwaitable)
-        {
-            ConfiguredTaskAwaitable<T>.ConfiguredTaskAwaiter awaiter = configuredAwaitable.GetAwaiter();
-            if (!awaiter.IsCompleted)
-            {
-                UnsafeAwaitAwaiterFromRuntimeAsync(awaiter);
-            }
-
-            return awaiter.GetResult();
-        }
-
-        // Marked intrinsic since JIT recognises the helper by name when doing optimizations.
-        [Intrinsic]
-        [BypassReadyToRun]
-        [MethodImpl(MethodImplOptions.Async)]
-        public static T Await<T>(ConfiguredValueTaskAwaitable<T> configuredAwaitable)
-        {
-            ConfiguredValueTaskAwaitable<T>.ConfiguredValueTaskAwaiter awaiter = configuredAwaitable.GetAwaiter();
-            if (!awaiter.IsCompleted)
-            {
-                UnsafeAwaitAwaiterFromRuntimeAsync(awaiter);
-            }
-
-            return awaiter.GetResult();
-        }
-#endif
     }
 }
